refactor: pair programming - use View class; add Campaign
refactor: pair programming - use View class; add Campaign
